feat(ci): EAS-97 pipelines for production release (#14)
This PR adds Github CI pipelines that trigger production builds for docker images on pull requests on the main branch. The built production images will be released to the Github package registry.

The images will only be tagged as latest since proper semver has not be clearly defined by the team and is still object of discussions. For now, though, latest tag will be more than fine since we want to quickly update the production instances without worrying too much about the versions.